refactor(graph): decompose audio graph validation, specs, and tests
