+++ src-tauri/src/audio/graph/mod.rs
@@ -1,0 +1,7 @@
+pub mod types;
+pub mod validate;
+
+pub use types::*;
+
+#[cfg(test)]
+mod tests;
